pyproject.toml
[project]
name = "rikai-slide-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "gradio>=6.2.0",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
